refactor: move download cluster to handler directory, rename to camelCase (ADR-0012 Option A)
Move downloadOrchestrator, failureHandler, s3Recovery from
src/services/download/ to src/lambdas/sqs/StartFileUpload/ (single
consumer). Rename kebab-case handler-local files to camelCase.
src/services/download/ retains only multi-consumer files.

Confidence: high
Scope-risk: moderate
