+++ src/lambdas/sqs/SendPushNotification/endpointCleanupHelpers.ts
