+++ src/lambdas/sqs/SendPushNotification/index.ts
@@ -10,8 +10,8 @@ import {addAnnotation, addMetadata, endSpan, logError, logInfo, metrics, MetricU
 import type {FileNotificationType} from '#types/notificationTypes'
 import {pushNotificationAttributesSchema} from '#types/schemas'
 import {validateSchema} from '@mantleframework/validation'
-import {cleanupDisabledEndpoints} from './endpoint-cleanup-helpers.js'
-import {getDevice, getDeviceIdsForUser, mapSettledToDeviceResults, processNotificationResults, sendNotificationToDevice} from './push-helpers.js'
+import {cleanupDisabledEndpoints} from './endpointCleanupHelpers.js'
+import {getDevice, getDeviceIdsForUser, mapSettledToDeviceResults, processNotificationResults, sendNotificationToDevice} from './pushHelpers.js'
 
 const sqs = defineSqsHandler<string>({operationName: 'SendPushNotification', parseBody: false, queue: 'SendPushNotification'})
