+++ src/lambdas/sqs/StartFileUpload/failureHandler.ts
