+++ src/lambdas/sqs/StartFileUpload/fileHelpers.ts
